prior to the merge, the mpm tree was tagged with apache-1_3-merge-2-pre,
and after the merge, the mpm tree will be tagged with
apache-1_3-merge-2-post


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@83709 13f79535-47bb-0310-9956-ffa450edef68

+++ os/unix/unixd.c
@@ -198,7 +198,8 @@ int unixd_setup_child(void)
 	if (setuid(unixd_config.user_id) == -1) {
 	    GETUSERMODE();
 	    ap_log_error(APLOG_MARK, APLOG_ALERT, NULL,
-			"setuid: unable to change to uid: %d", unixd_config.user_id);
+			"setuid: unable to change to uid: %ld",
+                        (long) unixd_config.user_id);
 	    exit(1);
 	}
 	GETUSERMODE();
@@ -211,7 +212,8 @@ int unixd_setup_child(void)
 #endif
 	setuid(unixd_config.user_id) == -1)) {
 	ap_log_error(APLOG_MARK, APLOG_ALERT, NULL,
-		    "setuid: unable to change to uid: %d", unixd_config.user_id);
+		    "setuid: unable to change to uid: %ld",
+                    (long) unixd_config.user_id);
 	return -1;
     }
 #endif
